Fix code not being editable after being run
Will need to do something about the history being clobbered up (which was the case with other way too but would have been easier slightly to fix). Looked into it a bit, and would be too complicated (and ambiguous) to try and edit the output, so I think the best option is to add an uneditable overlay with the output.

This would be something like a CodeEdit which has a totally transparent background and is kept in sync. Or perhaps simpler to just add pop-ups to it with the output which are kept in-line with it.